# src/test/resources/open-api-doc.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/audit-logs/": {
      "get": {
        "tags": [
          "audit-logs-controller"
        ],
        "operationId": "getAuditInfo",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditEvent"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuditEvent": {
        "type": "object",
        "properties": {
          "sourceObjId": {
            "type": "string"
          },
          "sourceObjectName": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "actionType": {
            "type": "string"
          },
          "timeStamp": {
            "type": "string",
            "format": "date-time"
          },
          "service": {
            "type": "string"
          },
          "instanceId": {
            "type": "string"
          }
        }
      }
    }
  }
}